Moved sorting function to hooks
