+++ plugins/scaffolder-react/src/next/hooks/index.ts
@@ -19,3 +19,4 @@ export {
   type ParsedTemplateSchema,
 } from './useTemplateSchema';
 export { useTemplateParameterSchema } from './useTemplateParameterSchema';
+export { useFeatureFlaggedProperties } from './useFeatureFlaggedProperties';
